requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-extensions==3.0.9
Faker==4.9.0
flake8==3.8.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.4
wcwidth==0.2.5
